refactor: move url function to url_utils.py
